pyproject.toml
[project]
name = "tiny-ai"
version = "0.1.0"
description = "Local RAG chatbot library, no API keys needed"
requires-python = ">=3.9"

dependencies = [
    "faiss-cpu",
    "sentence-transformers",
    "llama-cpp-python",
    "huggingface-hub",
    "pymupdf",
    "numpy",
]

[project.urls]
Homepage = "https://github.com/Madhan-1000/tiny-ai"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.backends.legacy:build"

[tool.setuptools.packages.find]
where = ["src"]